fix unit tests with active boolean

+++ internal/control/collective.go
@@ -245,7 +245,9 @@ func syncData() (err error) {
 
 		}(dataToStore)
 
-		client.SyncDataRequest(&controller.ReplicaNodes[0].IpAddress, dataToStore)
+		if active {
+			client.SyncDataRequest(&controller.ReplicaNodes[0].IpAddress, dataToStore)
+		}
 	}
 	return nil
 }
@@ -335,7 +337,9 @@ func removeNode(replicationGroup int) (nodeRemoved Node, err error) {
 				Database: data[i].Database,
 			})
 		}
-		client.DeleteData(&nodeRemoved.IpAddress, &protoData)
+		if active {
+			client.DeleteData(&nodeRemoved.IpAddress, &protoData)
+		}
 
 		// Dictionary update to remove from dictionary groups
 		// Create the data update request object
@@ -391,11 +395,15 @@ func removeNode(replicationGroup int) (nodeRemoved Node, err error) {
 			}
 		}
 		if len(dataToRemove) > 50 {
-			go callToRemove(dataToRemove)
+			if active {
+				go callToRemove(dataToRemove)
+			}
 			dataToRemove = []Data{}
 		}
 	}
-	go callToRemove(dataToRemove)
+	if active {
+		go callToRemove(dataToRemove)
+	}
 
 	return nodeRemoved, nil
 }
@@ -510,7 +518,6 @@ func distributeData(key, bucket *string, data *[]byte) error {
 	}
 
 	// TODO: API - send data update to ReplicaStoreData - ReplicaUpdate rpc
-	
 
 	// Add this node to the DataDictionary
 	updateType := addToDataDictionary(newData)
@@ -525,27 +532,29 @@ func distributeData(key, bucket *string, data *[]byte) error {
 	}
 	log.Println(updateData)
 
-	// Fire off DataDictionary update process through the collective - DictionaryUpdate rpc
-	// Create the data update request object
-	updateDictionary := make(chan *proto.DataUpdates)
+	if active {
+		// Fire off DataDictionary update process through the collective - DictionaryUpdate rpc
+		// Create the data update request object
+		updateDictionary := make(chan *proto.DataUpdates)
 
-	// Call the dictionary function before passing the data into the channel
-	// send the update to the first node in the list
-	client.DictionaryUpdate(&controller.Data.CollectiveNodes[0].ReplicaNodes[0].IpAddress, updateDictionary)
+		// Call the dictionary function before passing the data into the channel
+		// send the update to the first node in the list
+		client.DictionaryUpdate(&controller.Data.CollectiveNodes[0].ReplicaNodes[0].IpAddress, updateDictionary)
 
-	updateDictionary <- &proto.DataUpdates{
-		CollectiveUpdate: &proto.CollectiveDataUpdate{
-			Update:     true,
-			UpdateType: int32(updateType),
-			Data: &proto.CollectiveData{
-				ReplicaNodeGroup:  int32(newData.ReplicaNodeGroup),
-				DataKey:           newData.DataKey,
-				Database:          newData.Database,
-				ReplicatedNodeIds: newData.ReplicatedNodeIds,
+		updateDictionary <- &proto.DataUpdates{
+			CollectiveUpdate: &proto.CollectiveDataUpdate{
+				Update:     true,
+				UpdateType: int32(updateType),
+				Data: &proto.CollectiveData{
+					ReplicaNodeGroup:  int32(newData.ReplicaNodeGroup),
+					DataKey:           newData.DataKey,
+					Database:          newData.Database,
+					ReplicatedNodeIds: newData.ReplicatedNodeIds,
+				},
 			},
-		},
+		}
+		updateDictionary <- nil
 	}
-	updateDictionary <- nil
 
 	return nil
 }
